COVID-19 Utility (Application) v1.3
The first version of the COVID-19 Utility Application. You can run it on your IDE, from there you will need to input the corresponding U.S State/County data into your Console/Terminal on your computer, from there the COVID statistics will display on the IDE. This is the first version, and many more changes will be made, not all numbers are considered to be accurate, as this program is currently (W.I.P)

This is the first release of the Application version of the COVID-19 Utility, which is being updated alongside the Web-Application based version of the COVID Utility.