COVID-19 Utility (Application) v1.3.1
COVID-19 Utility (Application) v1.3.1
The first version of the COVID-19 Utility (Application). You can run it on your IDE.
This is the first version, and many more changes will be made, not all numbers are considered to be accurate, as this program is currently (W.I.P)

This is a minor patch for the first release of the program.

Changes:
- Renamed/Refactored statedata and countrydata files.
- Minor changes to code to allow for less major errors while in IDE.
- Minor additions to codebase regarding sorting statistics and reading JSON/CSV files
- Cleaned up codebase
- Minor edits and tweaks, general improvements

Issues:
Program has issues with circular imports/dependencies, will be fixed in a future update.