COVID-19 Utility (App) v1.5.2
COVID-19 Utility (App) v1.5.2

The first version of the COVID-19 Utility (App). You can run it on your IDE, which will require you to input the corresponding U.S State/County data into your Console/Terminal on your computer, from there the COVID statistics will display. This is the first version, and many more changes will be made, not all numbers are considered to be accurate, as this program is currently (W.I.P)

This is a minor patch for the second (Pre)release of the program.

Changes:
- Added CDC COVID-19 Data API (not yet usable)
- Cleaned up codebase *slightly*

Issues: (To be fixed)
- Missing API sources
- NYT county data is having issues retrieving data
- Missing statistics and infographics relating to COVID data 

Roadmap:
- Adding rest of NYT/JHU/CDC API terms and functionality for more COVID data
- Adding permissions for location access to be used for gathering COVID Data, as an alternative to selecting
- Updating selectbox to allow for state and county options to be used rather than typing in to console/terminal
- Making the webpage more visually interesting to look at
- Adding TwitterBot functionality (To Be Determined)
- Adding more COVID-19 Data statistics, such as ICU Bed Capacity and Community Levels (per CDC)
- COVID-19 Data Predictive Model (To Be Determined)